+++ analytics/tests/test_counts.py
@@ -66,7 +66,6 @@ from zerver.models import (
     Client,
     DirectMessageGroup,
     Message,
-    NamedUserGroup,
     PreregistrationUser,
     RealmAuditLog,
     Recipient,
@@ -75,7 +74,6 @@ from zerver.models import (
     UserProfile,
 )
 from zerver.models.clients import get_client
-from zerver.models.groups import SystemGroups
 from zerver.models.messages import Attachment
 from zerver.models.realm_audit_logs import AuditLogEventType
 from zerver.models.scheduled_jobs import NotificationTriggers
@@ -102,11 +100,6 @@ class AnalyticsTestCase(ZulipTestCase):
         super().setUp()
         self.default_realm = do_create_realm(
             string_id="realmtest", name="Realm Test", date_created=self.TIME_ZERO - 2 * self.DAY
-        )
-        self.administrators_user_group = NamedUserGroup.objects.get(
-            name=SystemGroups.ADMINISTRATORS,
-            realm=self.default_realm,
-            is_system_group=True,
         )
 
         # used to generate unique names in self.create_*