+++ zerver/lib/event_schema.py
@@ -64,6 +64,7 @@ group_setting_type = UnionType(
 # larger "subscription" events that also contain personal settings.
 default_stream_fields = [
     ("is_archived", bool),
+    ("can_administer_channel_group", group_setting_type),
     ("can_remove_subscribers_group", group_setting_type),
     ("creator_id", OptionalType(int)),
     ("date_created", int),