+++ api_docs/changelog.md
@@ -20,6 +20,20 @@ format used by the Zulip server that they are interacting with.
 
 ## Changes in Zulip 10.0
 
+**Feature level 325**
+
+* [`GET /users/me/subscriptions`](/api/get-subscriptions),
+  [`GET /streams`](/api/get-streams), [`GET /events`](/api/get-events),
+  [`POST /register`](/api/register-queue): Added `can_administer_channel_group`
+  which is a [group-setting value](/api/group-setting-values) describing the
+  set of users with permissions to administer the channel in addition to realm
+  admins.
+* [`POST /users/me/subscriptions`](/api/subscribe),
+  [`PATCH /streams/{stream_id}`](/api/update-stream): Added `can_administer_channel_group`
+  which is a [group-setting value](/api/group-setting-values) describing the
+  set of users with permissions to administer the channel in addition to realm
+  admins.
+
 **Feature level 324**
 
 * [`POST /register`](/api/register-queue), [`GET /events`](/api/get-events),